alfred/bindings/swap/requirements.txt
pyzmq==13.1.0